Restructure umbrella crate (#277)
The plan consists of restructuring Umbrella crate following best practices to simplify the understanding of each component.
