+++ huginn-net/src/parser/mod.rs
@@ -1,0 +1,3 @@
+pub mod packet;
+
+pub use packet::*;
